refactor the zxporter gpu exporter to zxporter nodemon
