--- helm-chart/zxporter-gpu-exporter/Chart.yaml
@@ -1,6 +1,0 @@
-apiVersion: v2
-name: zxporter-gpu-exporter
-description: A Helm chart for Kubernetes
-type: application
-version: 0.0.1
-appVersion: "0.0.1"
