--- helm-chart/zxporter-gpu-exporter/templates/dcgm-exporter-configmap.yaml
@@ -1,55 +1,0 @@
-{{ if .Values.dcgmExporter.enabled }}
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: {{- include "dcgm-exporter.config-map" . | indent 1 }}
-data:
-  counters.csv: |
-    # Temperature and power usage,,
-    DCGM_FI_DEV_GPU_TEMP, gauge, Current temperature readings for the device in degrees C.
-    DCGM_FI_DEV_MEMORY_TEMP, gauge, Memory temperature for the device.
-    DCGM_FI_DEV_POWER_USAGE, gauge, Power usage for the device in Watts.
-
-    # Utilization,,
-    # DCGM_FI_DEV_GPU_UTIL provides overall GPU utilization which is useful for scenarios
-    # like fractional GPU sharing (e.g., EKS time-slicing, MIG) where profiling metrics may not be available.
-    DCGM_FI_DEV_GPU_UTIL, gauge, GPU utilization (in %).
-    # DCGM_FI_PROF_SM_ACTIVE, gauge, The ratio of cycles an SM has at least 1 warp assigned
-    # DCGM_FI_PROF_SM_OCCUPANCY, gauge, The fraction of resident warps on a multiprocessor
-    # DCGM_FI_PROF_GR_ENGINE_ACTIVE, gauge, Ratio of time the graphics engine is active (in %).
-    # DCGM_FI_PROF_DRAM_ACTIVE, gauge, The ratio of cycles the device memory interface is active sending or receiving data.
-
-    # Memory usage,,
-    DCGM_FI_DEV_FB_FREE, gauge, Framebuffer memory free (in MiB).
-    DCGM_FI_DEV_FB_USED, gauge, Framebuffer memory used (in MiB).
-    DCGM_FI_DEV_FB_TOTAL, gauge, Total Frame Buffer of the GPU in MB.
-    DCGM_FI_DEV_MEM_COPY_UTIL, gauge, Utilization of the memory copy engine.
-
-    # PCIE,,
-    # DCGM_FI_PROF_PCIE_TX_BYTES, gauge, Total number of bytes transmitted through PCIe TX
-    # DCGM_FI_PROF_PCIE_RX_BYTES, gauge, Total number of bytes received through PCIe RX
-    DCGM_FI_DEV_PCIE_LINK_GEN, gauge, PCIe Current Link Generation.
-    DCGM_FI_DEV_PCIE_LINK_WIDTH, gauge, PCIe Current Link Width.
-
-    # Pipelines,,
-    # DCGM_FI_PROF_PIPE_INT_ACTIVE, gauge, Ratio of cycles the integer pipe is active.
-    # DCGM_FI_PROF_PIPE_FP16_ACTIVE, gauge, Ratio of cycles the fp16 pipe is active.
-    # DCGM_FI_PROF_PIPE_FP32_ACTIVE, gauge, Ratio of cycles the fp32 pipe is active.
-    # DCGM_FI_PROF_PIPE_FP64_ACTIVE, gauge, Ratio of cycles the fp64 pipe is active.
-    # DCGM_FI_PROF_PIPE_TENSOR_ACTIVE, gauge, The ratio of cycles the tensor (HMMA) pipe is active (off the peak sustained elapsed cycles)
-
-    # Health,,
-    # DCGM_FI_DEV_CLOCKS_EVENT_REASONS is not supported by DCGM 3.3.7
-    # DCGM_FI_DEV_CLOCKS_EVENT_REASONS, gauge, Current clock event reasons (bitmask of DCGM_CLOCKS_EVENT_REASON_*)
-    DCGM_FI_DEV_XID_ERRORS, gauge, The value is the specific XID error
-    DCGM_FI_DEV_POWER_VIOLATION, gauge, Power Violation time in ns.
-    DCGM_FI_DEV_THERMAL_VIOLATION, gauge, Thermal Violation time in ns.
-
-    # NVLink,,
-    # DCGM_FI_PROF_NVLINK_TX_BYTES, gauge, The number of bytes of active NvLink tx (transmit) data including both header and payload.
-    # DCGM_FI_PROF_NVLINK_RX_BYTES, gauge, The number of bytes of active NvLink rx (read) data including both header and payload.
-
-    # Clocks,,
-    DCGM_FI_DEV_SM_CLOCK, gauge, SM clock frequency (in MHz).
-    DCGM_FI_DEV_MEM_CLOCK, gauge, Memory clock frequency (in MHz).
-{{- end }}
