--- helm-chart/zxporter-gpu-exporter/templates/gpu-exporter-configmap.yaml
@@ -1,9 +1,0 @@
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: {{- include "zxporter-gpu-exporter.config-map" . | indent 1}}
-data:
-{{- $config := .Values.gpuMetricsExporter.config | default dict }}
-{{- if $config }}
-  {{- toYaml $config | nindent 2 }}
-{{- end }}
