--- helm-chart/zxporter-gpu-exporter/templates/rbac.yaml
@@ -1,72 +1,0 @@
-{{- if .Values.serviceAccount.create -}}
-apiVersion: v1
-kind: ServiceAccount
-metadata:
-  name: {{ include "zxporter-gpu-exporter.serviceAccountName" . }}
-  namespace: {{ .Release.Namespace }}
-  labels:
-    {{- include "zxporter-gpu-exporter.labels" . | nindent 4 }}
-  {{- with .Values.serviceAccount.annotations }}
-  annotations:
-    {{- toYaml . | nindent 4 }}
-  {{- end }}
-
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: {{ if .Values.gpuMetricsExporter.rbac.clusterWide }}ClusterRole{{ else }}Role{{ end }}
-metadata:
-  name: {{ include "zxporter-gpu-exporter.fullname" . }}
-  labels:
-    {{- include "zxporter-gpu-exporter.labels" . | nindent 4 }}
-  {{- if not .Values.gpuMetricsExporter.rbac.clusterWide }}
-  namespace: {{.Release.Namespace}}
-  {{- end }}
-rules:
-- apiGroups:
-  - ""
-  resources:
-  - pods
-  verbs:
-  - get
-  - list
-- apiGroups:
-    - apps
-  resources:
-    - replicasets
-    - deployments
-    - statefulsets
-    - daemonsets
-  verbs:
-    - get
-- apiGroups:
-    - batch
-  resources:
-    - jobs
-    - cronjobs
-  verbs:
-    - get
-- apiGroups:
-    - argoproj.io
-  resources:
-    - rollouts
-  verbs:
-    - get
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: {{ if .Values.gpuMetricsExporter.rbac.clusterWide }}ClusterRoleBinding{{ else }}RoleBinding{{ end }}
-metadata:
-  name: {{ include "zxporter-gpu-exporter.fullname" . }}
-  labels:
-    {{- include "zxporter-gpu-exporter.labels" . | nindent 4 }}
-  {{- if not .Values.gpuMetricsExporter.rbac.clusterWide }}
-  namespace: {{.Release.Namespace}}
-  {{- end }}
-roleRef:
-  apiGroup: rbac.authorization.k8s.io
-  kind: {{ if .Values.gpuMetricsExporter.rbac.clusterWide }}ClusterRole{{ else }}Role{{ end }}
-  name: {{ include "zxporter-gpu-exporter.fullname" . }}
-subjects:
-- kind: ServiceAccount
-  name: {{ include "zxporter-gpu-exporter.serviceAccountName" . }}
-  namespace: {{.Release.Namespace}}
-{{- end }}
