+++ helm-chart/zxporter-nodemon/Chart.yaml
@@ -1,0 +1,6 @@
+apiVersion: v2
+name: zxporter-nodemon
+description: A Helm chart for Kubernetes
+type: application
+version: 0.0.1
+appVersion: "0.0.1"
