+++ helm-chart/zxporter-nodemon/templates/nodemon-configmap.yaml
@@ -1,0 +1,9 @@
+apiVersion: v1
+kind: ConfigMap
+metadata:
+  name: {{- include "zxporter-nodemon.config-map" . | indent 1}}
+data:
+{{- $config := .Values.gpuMetricsExporter.config | default dict }}
+{{- if $config }}
+  {{- toYaml $config | nindent 2 }}
+{{- end }}
